+++ packages/plugins/auto-doc/src/pdf/rendering/render-png-from-html.ts
@@ -14,7 +14,7 @@ type ExportPNGProps = ExportProps & {
   quality?: number;
 };
 
-export async function exportPngFromHTMLElement({
+export async function renderPngFromHtml({
   element,
   imageWidth,
   imageHeight,